+++ apps/synt-vedtakshistorikk-service/src/test/java/no/nav/testnav/apps/syntvedtakshistorikkservice/provider/BrukereControllerTest.java
@@ -1,12 +1,9 @@
 package no.nav.testnav.apps.syntvedtakshistorikkservice.provider;
 
-import no.nav.testnav.apps.syntvedtakshistorikkservice.consumer.PdlProxyConsumer;
 import no.nav.testnav.apps.syntvedtakshistorikkservice.provider.request.SyntetiserArenaRequest;
-import no.nav.testnav.apps.syntvedtakshistorikkservice.service.ArenaForvalterService;
-import no.nav.testnav.apps.syntvedtakshistorikkservice.service.IdentService;
+import no.nav.testnav.apps.syntvedtakshistorikkservice.service.BrukerService;
 import no.nav.testnav.libs.domain.dto.arena.testnorge.brukere.Arbeidsoeker;
 import no.nav.testnav.libs.domain.dto.arena.testnorge.vedtak.NyeBrukereResponse;
-import no.nav.testnav.libs.dto.personsearchservice.v1.PersonDTO;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -23,25 +20,16 @@ import java.util.Map;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.Mockito.when;
-import static no.nav.testnav.apps.syntvedtakshistorikkservice.service.util.ServiceUtils.MAKSIMUM_ALDER;
-import static no.nav.testnav.apps.syntvedtakshistorikkservice.service.util.ServiceUtils.MINIMUM_ALDER;
-import static no.nav.testnav.apps.syntvedtakshistorikkservice.service.VedtakshistorikkService.SYNT_TAGS;
 
 
 @RunWith(MockitoJUnitRunner.class)
 public class BrukereControllerTest {
 
     @Mock
-    private ArenaForvalterService arenaForvalterService;
-
-    @Mock
-    private IdentService identService;
-
-    @Mock
-    private PdlProxyConsumer pdlProxyConsumer;
+    private BrukerService brukerService;
 
     @InjectMocks
-    private BrukereController brukereController;
+    private BrukerController brukerController;
 
     private SyntetiserArenaRequest syntetiserArenaRequestSingle;
 
@@ -65,16 +53,9 @@ public class BrukereControllerTest {
 
     @Test
     public void registrerAntallIdenterMedOppfoelgingIArenaForvalter() {
-        var identer = Collections.singletonList(fnr1);
-        var personer = Collections.singletonList(PersonDTO.builder().ident(fnr1).build());
+        when(brukerService.registrerArenaBrukereMedOppfoelging(syntetiserArenaRequestSingle)).thenReturn(oppfoelgingResponse);
 
-        when(identService.getUtvalgteIdenterIAldersgruppe(1, MINIMUM_ALDER, MAKSIMUM_ALDER, null)).thenReturn(personer);
-        when(pdlProxyConsumer.createTags(identer, SYNT_TAGS)).thenReturn(true);
-        when(arenaForvalterService
-                .opprettArbeidssoekereUtenVedtak(identer, miljoe))
-                .thenReturn(oppfoelgingResponse);
-
-        ResponseEntity<Map<String, NyeBrukereResponse>> result = brukereController.registrerBrukereIArenaForvalterMedOppfoelging(syntetiserArenaRequestSingle);
+        ResponseEntity<Map<String, NyeBrukereResponse>> result = brukerController.registrerBrukereIArenaForvalterMedOppfoelging(syntetiserArenaRequestSingle);
         assertThat(result.getBody().keySet()).hasSize(1);
         assertThat(result.getBody()).containsKey(fnr1);
         assertThat(result.getBody().get(fnr1).getArbeidsoekerList().get(0).getPersonident()).isEqualTo(fnr1);
@@ -85,7 +66,7 @@ public class BrukereControllerTest {
     public void checkExceptionOccursOnBadMiljoe() {
         var request = new SyntetiserArenaRequest("test", 1);
         assertThrows(ResponseStatusException.class, () -> {
-            brukereController.registrerBrukereIArenaForvalterMedOppfoelging(request);
+            brukerController.registrerBrukereIArenaForvalterMedOppfoelging(request);
         });
     }
 }
